Create Lib Management
Library Management System is a software used to manages the catalog of a library. This helps to keep the records of whole transactions of the books available in the library. Library Management System is very easy to use and fulfills all the requirement of a librarian. There are many features which helps librarian to keep records of available books as well as issued books.